move tests of DisplayRange to testsFigure
